+++ CHANGELOG.md
@@ -5,9 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-  - [447](https://github.com/thoth-pub/thoth/issues/447) - Google Books output must contain at least one BISAC subject code
-  - [404](https://github.com/thoth-pub/thoth/issues/404) - JSTOR output must contain at least one BISAC subject code
+### Changed
   - [499](https://github.com/thoth-pub/thoth/issues/499) - Default main\_contribution to true
+
+### Fixed
+  - [447](https://github.com/thoth-pub/thoth/issues/447) - Prevents Google Books Onix3 format output from Export API if Thoth record doesn't contain at least one BIC or BISAC subject code
+  - [404](https://github.com/thoth-pub/thoth/issues/404) - Prevents JSTOR Onix3 format output from Export API if Thoth record doesn't contain at least one BIC or BISAC subject code
 
 ## [[0.11.16]](https://github.com/thoth-pub/thoth/releases/tag/v0.11.16) - 2024-02-19
 ### Changed